+++ src/usecases/budget.rs
@@ -3,22 +3,22 @@ use rusqlite::{Connection, Result};
 pub fn budget_total_equal(
     conn: &Connection,
     category: &str,
-) -> Result<(u64, u64, u64, bool), rusqlite::Error> {
+) -> Result<(f64, f64, f64, bool), rusqlite::Error> {
     let mut stmt = conn.prepare("select total_amount from totalamount")?;
-    let total_amount: u64 = stmt.query_row([], |row| row.get::<_, u64>(0))?;
+    let total_amount: f64 = stmt.query_row([], |row| row.get::<_, f64>(0))?;
 
     let mut stmt = conn.prepare("select sum(amount) from budget")?;
-    let budget_total_sum: u64 = stmt
-        .query_row([], |row| row.get::<_, Option<u64>>(0))?
-        .unwrap_or(0);
+    let budget_total_sum: f64 = stmt
+        .query_row([], |row| row.get::<_, Option<f64>>(0))?
+        .unwrap_or(0.0);
 
     let mut stmt = conn.prepare("select sum(amount) from budget where category <> ?")?;
-    let budget_except_sum: Option<u64> = stmt.query_row([category], |row| row.get(0))?;
+    let budget_except_sum: Option<f64> = stmt.query_row([category], |row| row.get(0))?;
 
     Ok((
         total_amount,
         budget_total_sum,
-        budget_except_sum.unwrap_or(0),
+        budget_except_sum.unwrap_or(0.0),
         budget_total_sum < total_amount,
     ))
 }
@@ -35,19 +35,19 @@ pub fn budget_data_exists(conn: &Connection) -> Result<bool> {
     Ok(exists)
 }
 
-pub fn budget_amount(conn: &Connection, category: &str) -> Result<(u64, u64), rusqlite::Error> {
+pub fn budget_amount(conn: &Connection, category: &str) -> Result<(f64, f64), rusqlite::Error> {
     let mut stmt =
         conn.prepare("select amount, remaining_amount from budget where category = ?")?;
-    let amount = stmt.query_row([category], |row| row.get::<_, u64>(0))?;
-    let remaining_amount = stmt.query_row([category], |row| row.get::<_, u64>(1))?;
+    let amount = stmt.query_row([category], |row| row.get::<_, f64>(0))?;
+    let remaining_amount = stmt.query_row([category], |row| row.get::<_, f64>(1))?;
     Ok((amount, remaining_amount))
 }
 
 pub fn calculate_budget(
     conn: &Connection,
     category: &str,
-    spending_amount: u64,
-    spending_sum_category: u64,
+    spending_amount: f64,
+    spending_sum_category: f64,
 ) {
     let (_, remaining) = budget_amount(conn, category).unwrap_or_else(|e| panic!("Err: {}", e));
     let remaining_amount = remaining - spending_amount;
